Update page metadata and remove outdated content
Removed a blog post and related links from various pages that are no longer relevant. Also, updated the Hugo generator version in the metadata of a page. These changes keep the webpage current and reduce clutter. The removed post and related links had outdated data which was not beneficial to keep in the site anymore.